css garbage, better nav links
