+++ public/styles.css
@@ -1,0 +1,8 @@
+body {
+  font-family: 'Quicksand', sans-serif;
+}
+
+a {
+  color: inherit;
+  text-decoration: inherit;
+}
